Former-commit-id: 766d660aef40e8c1e5fa8574fe43e8cdb74d0b2b

+++ core/service/dps/personfinance_service.go
@@ -32,19 +32,19 @@ func NewPersonFinanceService(rep personfinance.IPersonFinanceRepository,
 	}
 }
 
-func (p *personFinanceService) GetRiseInfo(personId int) (
+func (p *personFinanceService) GetRiseInfo(personId int32) (
 	personfinance.RiseInfoValue, error) {
 	pf := p._rep.GetPersonFinance(personId)
 	return pf.GetRiseInfo().Value()
 }
 
 // 开通增利服务
-func (p *personFinanceService) OpenRiseService(personId int) error {
+func (p *personFinanceService) OpenRiseService(personId int32) error {
 	m := p._accRep.GetMember(personId)
 	if m == nil {
 		return member.ErrNoSuchMember
 	}
-	if m.GetValue().Level < variable.PersonFinanceMinLevelLimit {
+	if m.GetValue().Level < int32(variable.PersonFinanceMinLevelLimit) {
 		return errors.New("会员等级不够,请升级后再开通理财账户！")
 	}
 	pf := p._rep.GetPersonFinance(personId)
@@ -52,7 +52,7 @@ func (p *personFinanceService) OpenRiseService(personId int) error {
 }
 
 // 提交转入/转出日志
-func (p *personFinanceService) CommitTransfer(personId, logId int) error {
+func (p *personFinanceService) CommitTransfer(personId, logId int32) error {
 	pf := p._rep.GetPersonFinance(personId)
 	rs := pf.GetRiseInfo()
 	if rs == nil {
@@ -62,7 +62,7 @@ func (p *personFinanceService) CommitTransfer(personId, logId int) error {
 }
 
 // 转入(业务放在service,是为person_finance解耦)
-func (p *personFinanceService) RiseTransferIn(personId int,
+func (p *personFinanceService) RiseTransferIn(personId int32,
 	transferWith personfinance.TransferWith, amount float32) (err error) {
 	pf := p._rep.GetPersonFinance(personId)
 	r := pf.GetRiseInfo()
@@ -106,7 +106,7 @@ func (p *personFinanceService) RiseTransferIn(personId int,
 }
 
 // 转出
-func (p *personFinanceService) RiseTransferOut(personId int,
+func (p *personFinanceService) RiseTransferOut(personId int32,
 	transferWith personfinance.TransferWith, amount float32) (err error) {
 	pf := p._rep.GetPersonFinance(personId)
 	r := pf.GetRiseInfo()
@@ -145,7 +145,7 @@ func (p *personFinanceService) RiseTransferOut(personId int,
 }
 
 // 结算收益(按日期每天结息)
-func (p *personFinanceService) RiseSettleByDay(personId int,
+func (p *personFinanceService) RiseSettleByDay(personId int32,
 	settleUnix int64, dayRatio float32) (err error) {
 	pf := p._rep.GetPersonFinance(personId)
 	r := pf.GetRiseInfo()
